backend/requirements.txt
asgiref==3.8.1
attrs==23.2.0
autoflake==2.3.1
cffi==1.16.0
cryptography==42.0.8
dataclass-type-validator==0.1.2
dataclasses==0.6
Django==4.1.13
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
drf-spectacular==0.27.2
et-xmlfile==1.1.0
inflection==0.5.1
isort==5.13.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
numpy==2.0.0
openpyxl==3.1.4
pandas==2.2.2
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.8.0
pymongo==3.12.3
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
six==1.16.0
sqlparse==0.2.4
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
